server/requirements.txt
Flask
Flask-Cors
Flask-Caching
Flask-Limiter
python-dotenv
requests
